Probleme integriert!

+++ app/controllers/solutions_controller.rb
@@ -3,6 +3,7 @@ class SolutionsController < ApplicationController
 
 
   def index
+    if File.exist?("HPPLAN_v1_OFV.txt")
       fi=File.open("HPPLAN_v1_OFV.txt", "r")
       line=fi.readline
       fi.close
@@ -10,6 +11,8 @@ class SolutionsController < ApplicationController
         if sa[1]!=nil                                                 #Wenn Zielfktwert vorhanden
         @objective_function_value=sa[1]                               #speichere Wert zur Ausgabe
         end
+    else
+    end
   end
 
 end
